added genoprobs and map arguments back for more flexibility
I was just trying to convert a DO object from the Churchill lab to cape, and I needed to supply the genoprobs and map separately. I imagine I won't be the only one to need that flexibility. The defaults are NULL, so if people just have a regular cross object, they won't need to supply genoprobs or map.