added invariant marker handling to genome.wide.threshold
added ability to handle invariant markers
These don't return models with the same dimensions, so I added a check for the dimensions of the model coefficients. If they only have one row, the function returns NA.
This shouldn't happen because we are supposed to do a check for invariant markers before we run cape, so I'm not sure how those snuck through. It's worth checking out.