fixed dimnames
pairscan.kin was pulling dimnames for the results matrices from the wrong place.
I added colnames for the results matrices to match the ones from pairscan.noKin

I also replaced a for loop with lapply to speed things up a bit. I tested it and went from 12 seconds to 10 seconds on about 3000 marker pairs