using saved data
The arguments of run.singlescan, run.pairscan, etc. only create confusion. I used these before as a way to decide for individual elements whether to reuse data or not. Now that we have a single parameter for re-using data, it doesn't make sense to have these arguments.
I have removed them, and replaced checks for whether to run individual steps based on whether the results exist and whether the user has asked to reuse data.
I also fixed a typo in the error.prop documentation