transform.to.phenospace
Transform to phenospace should be moved to the parameter file. This used to be set automatically to TRUE if people ran eigentraits, but sometimes people might want to have direct influences in terms of eigentraits. For now I have put it in the arguments with default to TRUE, but we should probably put it in the parameter file instead. We also need to add a check for scan.what. If people are scanning raw.traits or normalized traits instead of eigentraits, transform.to.phenospace must be FALSE