changes to pheno2covar
This function looked weird to me. It still had some naming conventions left over from an older version of cape, and it didn't seem to be behaving properly. I have updated the function to be more compatible with the new version of cape, but we will need to test its performance