Update get_marker_covar.R
The code was aligning individuals in the genotype matrix and the covariate matrix. If there was no covariate matrix, this resulted in empty marker data.

I changed this to align individuals in the phenotype and genotype matrices, and only look for covariates if one of the requested variables cannot be found in the genotype matrix