Update linkage_blocks_network.R
this choked if one there were any markers that couldn't be found in the genotype object. There is still something weird where it seems to enter the chromosome loop even for a covariate and I get an error for the cbind line about number of elements not being compatible. This line should be run for a covariate, and when I run the code line by line it's not run, so I'm not sure what's going on.